Create the policy allocation object
Will be the interface between the object mapper and the block allocator.
Block allocator logic would consist of choosing nodes to allocate from,
while the policy allocation would be aware of tiering and would provide
the block allocator the information of which nodes to choose from.
